feat(checklist): move application activation to registration
The ApplicationActivation got moved from Checklist.Library into a separat ApplicationActivation.Library Module.

Refs: CPLP-1986
