+++ src/Portal.Backend.sln
@@ -158,6 +158,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SdFactory.Library", "extern
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SdFactory.Library.Tests", "..\tests\externalsystems\SdFactory.Library.Tests\SdFactory.Library.Tests.csproj", "{A4BF3A98-BE55-4816-991F-086DC3B9E5FB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApplicationActivation.Library", "registration\ApplicationActivation.Library\ApplicationActivation.Library.csproj", "{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApplicationActivation.Library.Tests", "..\tests\registration\ApplicationActivation.Library.Tests\ApplicationActivation.Library.Tests.csproj", "{96D96CA7-35C0-40C6-A8C8-91E0C4456660}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -948,6 +952,30 @@ Global
 		{A4BF3A98-BE55-4816-991F-086DC3B9E5FB}.Release|x64.Build.0 = Release|Any CPU
 		{A4BF3A98-BE55-4816-991F-086DC3B9E5FB}.Release|x86.ActiveCfg = Release|Any CPU
 		{A4BF3A98-BE55-4816-991F-086DC3B9E5FB}.Release|x86.Build.0 = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|x64.Build.0 = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Debug|x86.Build.0 = Debug|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|x64.ActiveCfg = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|x64.Build.0 = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|x86.ActiveCfg = Release|Any CPU
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F}.Release|x86.Build.0 = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|x64.Build.0 = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Debug|x86.Build.0 = Debug|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|x64.ActiveCfg = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|x64.Build.0 = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|x86.ActiveCfg = Release|Any CPU
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1021,5 +1049,7 @@ Global
 		{4F0E9174-5DB5-42C5-9AA2-4914BDD29AE6} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
 		{7D3D6277-6E91-4F66-AB07-5049A66B5B3D} = {C8957230-4203-452C-A085-34091C5E370B}
 		{A4BF3A98-BE55-4816-991F-086DC3B9E5FB} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
+		{0BFA8C5B-3F15-4381-9621-AD7A3EC0315F} = {AB9C5AA2-DD5D-4A38-97C0-674A995C0AE0}
+		{96D96CA7-35C0-40C6-A8C8-91E0C4456660} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
 	EndGlobalSection
 EndGlobal
