+++ src/checklist/Checklist.Worker/Checklist.Worker.csproj
@@ -47,6 +47,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\..\framework\Framework.Async\Framework.Async.csproj" />
+        <ProjectReference Include="..\..\registration\ApplicationActivation.Library\ApplicationActivation.Library.csproj" />
         <ProjectReference Include="..\Checklist.Library\Checklist.Library.csproj" />
     </ItemGroup>
