+++ src/registration/ApplicationActivation.Library/ApplicationActivation.Library.csproj
@@ -1,0 +1,43 @@
+<!--
+- Copyright (c) 2021,2022 BMW Group AG
+- Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+-
+- See the NOTICE file(s) distributed with this work for additional
+- information regarding copyright ownership.
+-
+- This program and the accompanying materials are made available under the
+- terms of the Apache License, Version 2.0 which is available at
+- https://www.apache.org/licenses/LICENSE-2.0.
+-
+- Unless required by applicable law or agreed to in writing, software
+- distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+- WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+- License for the specific language governing permissions and limitations
+- under the License.
+-
+- SPDX-License-Identifier: Apache-2.0
+-->
+
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <AssemblyName>Org.Eclipse.TractusX.Portal.Backend.ApplicationActivation.Library</AssemblyName>
+        <RootNamespace>Org.Eclipse.TractusX.Portal.Backend.ApplicationActivation.Library</RootNamespace>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <UserSecretsId>91356e1f-3558-409f-9afe-a1cd06ae65b2</UserSecretsId>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\mailing\Mailing.SendMail\Mailing.SendMail.csproj" />
+      <ProjectReference Include="..\..\notifications\Notifications.Library\Notifications.Library.csproj" />
+      <ProjectReference Include="..\..\provisioning\Provisioning.Library\Provisioning.Library.csproj" />
+    </ItemGroup>
+
+</Project>
