+++ tests/registration/ApplicationActivation.Library.Tests/ApplicationActivation.Library.Tests.csproj
@@ -1,0 +1,49 @@
+<!--
+- Copyright (c) 2021,2022 BMW Group AG
+- Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+-
+- See the NOTICE file(s) distributed with this work for additional
+- information regarding copyright ownership.
+-
+- This program and the accompanying materials are made available under the
+- terms of the Apache License, Version 2.0 which is available at
+- https://www.apache.org/licenses/LICENSE-2.0.
+-
+- Unless required by applicable law or agreed to in writing, software
+- distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+- WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+- License for the specific language governing permissions and limitations
+- under the License.
+-
+- SPDX-License-Identifier: Apache-2.0
+-->
+
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsPackable>false</IsPackable>
+        <RootNamespace>Org.Eclipse.TractusX.Portal.Backend.ApplicationActivation.Library.Tests</RootNamespace>
+        <AssemblyName>Org.Eclipse.TractusX.Portal.Backend.ApplicationActivation.Library.Tests</AssemblyName>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="AutoFixture.AutoFakeItEasy" Version="4.17.0" />
+        <PackageReference Include="FakeItEasy" Version="7.3.1" />
+        <PackageReference Include="FluentAssertions" Version="6.7.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+        <PackageReference Include="xunit" Version="2.4.1" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="coverlet.collector" Version="3.1.2">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\..\..\src\registration\ApplicationActivation.Library\ApplicationActivation.Library.csproj" />
+        <ProjectReference Include="..\..\shared\Tests.Shared\Tests.Shared.csproj" />
+    </ItemGroup>
+</Project>
